requirements.txt
python-social-auth
django-wysiwyg
